+++ tests/cpp/test_resize.cpp
@@ -17,7 +17,7 @@
 #include <runtime/executor_utils.h>
 #include <runtime/fusion_executor_cache.h>
 #include <scheduler/tools/inlining.h>
-#include <scheduler/utils.h>
+#include <scheduler/tools/loop_domain_scheduler.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
 
@@ -3696,7 +3696,7 @@ TEST_F(ResizeTest, SliceScheduledLikeProducer) {
   fusion.addOutput(tv2);
 
   std::vector<IterDomain*> ref_loop = tv0->getLogicalDomain();
-  scheduler_utils::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
+  scheduler_tools::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
 
   for (auto tv : {tv1, tv2}) {
     tv->split(0, 32);
@@ -3744,7 +3744,7 @@ TEST_F(ResizeTest, PadScheduledLikeConsumer) {
   fusion.addOutput(tv3);
 
   std::vector<IterDomain*> ref_loop = tv2->getLogicalDomain();
-  scheduler_utils::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
+  scheduler_tools::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
 
   for (auto tv : {tv1, tv2, tv3}) {
     tv->split(0, 32);
@@ -3796,7 +3796,7 @@ TEST_F(ResizeTest, SliceThenPadLeftHalf) {
   fusion.addOutput(tv3);
 
   std::vector<IterDomain*> ref_loop = tv0->getLogicalDomain();
-  scheduler_utils::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
+  scheduler_tools::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
 
   for (auto tv : {tv1, tv2, tv3}) {
     tv->split(0, 32);
@@ -3851,7 +3851,7 @@ TEST_F(ResizeTest, SliceThenPadRightHalf) {
   fusion.addOutput(tv3);
 
   std::vector<IterDomain*> ref_loop = tv0->getLogicalDomain();
-  scheduler_utils::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
+  scheduler_tools::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
 
   for (auto tv : {tv1, tv2, tv3}) {
     tv->split(0, 32);
@@ -3915,7 +3915,7 @@ TEST_F(ResizeTest, SliceThenConcat) {
   fusion.addOutput(tv6);
 
   std::vector<IterDomain*> ref_loop = tv0->getLogicalDomain();
-  scheduler_utils::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
+  scheduler_tools::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
 
   for (auto tv : {tv1, tv2, tv3, tv4, tv5, tv6}) {
     tv->split(0, 32);
@@ -4002,7 +4002,7 @@ TEST_F(ResizeTest, SliceSliceConcatConcat) {
   fusion.addOutput(tv13);
 
   std::vector<IterDomain*> ref_loop = tv0->getLogicalDomain();
-  scheduler_utils::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
+  scheduler_tools::scheduleLoopDomainsLike(fusion.allTvs(), ref_loop);
 
   for (auto tv : fusion.allTvs()) {
     if (tv->isFusionInput()) {
